requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
SQLAlchemy==1.4.46
Flask-SQLAlchemy==2.5.1
pymysql==1.0.3
google-auth==2.17.3
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
cryptography==40.0.2
cachetools==5.3.0